Add files via upload
Given the following class hierarchy



add at least 3 classes in the third level of the hierarchy (which do not all inherit from the same class). Each of these added classes will have their own attribute, that is, not inherited, as well as a drawing (not an image) of the tool in question. Note: if you consider it necessary, you can add more methods and / or include parameters.

Once the hierarchy is defined, develop a visual application that instants at least 4 tools of the third level (that is, of the classes that you added). The values ​​of the attributes will be assigned at design time, except those of the base class, which will be requested to the user when executing the application. Then you must do the following:

Show in the main frame all the information of each tool, one by one, together with its drawing, ordered descending by length of the description.
Show in another frame a "parade" of the drawings of all the tools in movement (horizontal, vertical, diagonal or inverted diagonal) ordered by key; All the attributes of each object must also be seen.
The following are the requirements to which the design of the application must be subject:

The values ​​of the attributes will have to be shown through a single method (which can not be toString ()).
The instanceOf operator must not be used.
The drawing and movement must be polymorphic methods in the class hierarchy, they must NOT be encoded in the frames.
The drawings must include at least the four basic types (line, rectangle, oval and String), and at least two different colors.
Include a button to assign key and example description for each created object.
Include an image of the class diagram of the designed hierarchy (it can be generated in NetBeans with a plugin).